Feature 28 create user controller and service (#32)
* Add DTO class CreateUserDto with validation

* Move CreateUserDto class file to correct directory

* Update .eslintrc.js

added prettier definition to clear all the red from vs code

* Update .eslintrc.js

added routes to create a user and delete a user

* updated for creator update

Added the route to update a user to creator.

* Update auth.controller.ts

* Update auth.service.ts

* Updates

finished renaming the userAuth.model so it didn't break anything. As well as enabling the UserModel, also added the password prop to user.model to stop it breaking functionality

* Refactor Admin routes

Removed admin routes from auth routes and added a working adminAddUser route to the user side. Also added a password to the user.module so that the admin add user route works.

* Added Delete route

Added a delete route for admins.

* Updated with admin update route

Further updated the user.control and user.services for an adminUpdate route

* Create user.controller.spec.ts

added mocked user tests for admin routes.

* Refactored to use createuserdto

I refactored the admin code to use the create user dto and removed the test cases cause I can't get the relative path working.

* More refactoring of the admin routes

updated the service and controller to use the CreateUserDto.

* Update activity.service.spec.ts

* Update activity.service.spec.ts

* Update activity.service.spec.ts

codefactor thinks this needs to be changed

* Updated names

Updated route names and removed extra routes

* Update user.controller.ts

* [CodeFactor] Apply fixes

* fix: renamed UserSchema to UserAuthSchema

* fix: adjusted endpoints to follow standards in wiki.

---------

Co-authored-by: Ali Sharif <alisharif29@gmail.com>
Co-authored-by: codefactor-io <support@codefactor.io>
Co-authored-by: theGaryLarson <gary.larson@seattlecolleges.edu>